pyproject.toml
[project]
name = "langgenie"
version = "0.1.0"
description = ""
authors = [
    {name = "amulyaprasanth",email = "amulyaprasanth301@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "chainlit (>=2.4.201,<3.0.0)",
    "langchain (>=0.3.21,<0.4.0)",
    "fastembed (>=0.6.0,<0.7.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "arxiv (>=2.1.3,<3.0.0)",
    "pymupdf (>=1.25.4,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-groq (>=0.3.1,<0.4.0)",
    "langgraph (>=0.3.18,<0.4.0)"
]

[tool.poetry]
packages = [{include = "langgenie", from = "src"}]


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
